backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.6
requests==2.32.3
beautifulsoup4==4.13.3
rapidfuzz==3.12.1
reportlab==4.2.5
